refactor all the so dashbord page do the fetching
